+++ src/main/features/digital-outcomes/controller/index.ts
@@ -2,10 +2,12 @@ import {
   GET_OPPORTUNITIES,
   GET_OPPORTUNITIES_DETAILS,
   GET_OPPORTUNITIES_API,
-  GET_OPPORTUNITIES_DETAILS_REVIE_RECOMMENDATION,
-  GET_OPPORTUNITIES_QA,
   OPPORTUNITIES_DOWNLOAD,
 } from './opportunities';
+
+import { GET_OPPORTUNITIES_DETAILS_REVIE_RECOMMENDATION, OPPORTUNITY_DETAILS_DOWNLOAD } from './opportunitiesDetails';
+
+import { GET_OPPORTUNITIES_QA } from './opportunitiesQA';
 
 export const DIGITALOUTCOMES_CONTROLLER = {
   /**
@@ -17,6 +19,7 @@ export const DIGITALOUTCOMES_CONTROLLER = {
   GET_OPPORTUNITIES_DETAILS_REVIE_RECOMMENDATION,
   GET_OPPORTUNITIES_QA,
   OPPORTUNITIES_DOWNLOAD,
+  OPPORTUNITY_DETAILS_DOWNLOAD,
   /**
    * @POST_VIEW
    */
