+++ src/main/features/digital-outcomes/controller/opportunities.ts
@@ -4,9 +4,6 @@ import * as fileData from '../../../resources/content/digital-outcomes/oppertuni
 // import * as procdata from '../../../resources/content/digital-outcomes/procdetails.json';
 import { TenderApi } from '../../../common/util/fetch/tenderService/tenderApiInstance';
 
-import * as inboxData from '../../../resources/content/event-management/qa.json';
-
-import moment from 'moment-business-days';
 import { DynamicFrameworkInstance } from 'main/features/event-management/util/fetch/dyanmicframeworkInstance';
 
 export const GET_OPPORTUNITIES = async (req: express.Request, res: express.Response) => {
@@ -268,138 +265,6 @@ export const GET_OPPORTUNITIES_DETAILS = async (req: express.Request, res: expre
     res.render('opportunitiesReview', display_fetch_data);
   } catch (error) {
     console.log('error in opportunities', error);
-  }
-};
-
-export const GET_OPPORTUNITIES_DETAILS_REVIE_RECOMMENDATION = async (req: express.Request, res: express.Response) => {
-  try {
-    const { projectId, status, subStatus, lot } = req.query;
-
-    //const eventTypeURL = 'https://dev-ccs-scale-cat-service.london.cloudapps.digital/tenders/projects/21737';
-    //let projectIds = '22111';
-    //let projectIds = '21737';
-
-    const eventTypeURL = `/tenders/projects/${projectId}`;
-
-    let getOppertunitiesData = await TenderApi.InstanceSupplierQA().get(eventTypeURL);
-    let getOppertunities = getOppertunitiesData?.data;
-    let awards: any = '';
-    let awardDate = '';
-    let awardsupplierId = '';
-    let part;
-    let award_matched_selector = [];
-    if (getOppertunities.records[0].compiledRelease?.awards) {
-      awards = getOppertunities.records[0].compiledRelease?.awards[0];
-      awardDate = moment(getOppertunities.records[0].compiledRelease?.awards[0]?.date).format('DD/MM/YYYY');
-      awardsupplierId = awards.suppliers[0].id;
-      part = getOppertunities.records[0].compiledRelease.parties;
-      //GB-COH-02299747
-
-      award_matched_selector = part?.filter((agroupitem: any) => {
-        //console.log('agroupitem?.id', agroupitem?.id);
-        return agroupitem?.id === awardsupplierId;
-      });
-
-      award_matched_selector = award_matched_selector[0];
-    }
-
-    let bids = getOppertunities.records[0].compiledRelease.bids;
-
-    let supplierSummeryCount = bids.map((item: any) => {
-      let result: any[] = [];
-      //   const newItem = item;
-      item.statistics.map((sta: any) => {
-        let measure = sta.measure;
-        let value = sta.value;
-        result[measure] = value;
-      });
-      return result;
-    });
-
-    const display_fetch_data = {
-      buyer: getOppertunities.records[0].compiledRelease.buyer,
-      tenderer: getOppertunities.records[0].compiledRelease.tender,
-      tenderers: getOppertunities.records[0].compiledRelease.tender.tenderers,
-      //parties: getOppertunities.records[0].compiledRelease.parties[0],
-      parties: award_matched_selector,
-
-      awards: awards,
-      awardDate: awardDate,
-      //endDate: moment(getOppertunities.records[0].compiledRelease.tenderPeriod.endDate).format('dddd DD MMMM YYYY'),
-      endDate: moment(getOppertunities.records[0].compiledRelease.tender.tenderPeriod.endDate).format(
-        'dddd DD MMMM YYYY'
-      ),
-      supplierSummeryCount: supplierSummeryCount,
-      suppliersResponded: supplierSummeryCount[0]['suppliersResponded'],
-      suppliersNotResponded: supplierSummeryCount[0]['suppliersNotResponded'],
-      ocid: getOppertunities.records[0].compiledRelease.ocid,
-      projectId: projectId,
-      status: status,
-      subStatus: subStatus,
-      currentLot: lot,
-    };
-
-    res.render('opportunitiesDetails', display_fetch_data);
-  } catch (error) {}
-};
-
-export const GET_OPPORTUNITIES_QA = async (req: express.Request, res: express.Response) => {
-  let appendData: any;
-  let eventIds: any;
-  let projectIds: any;
-  let isSupplierQA = false;
-  let projectId;
-  try {
-    if (req.query.id != undefined) {
-      eventIds = req.query.id;
-      projectIds = req.query.prId;
-      isSupplierQA = true;
-      projectId = req.query.prId;
-    } else {
-      eventIds = req.session.eventId;
-      projectIds = req.session.projectId;
-      projectId = req.session.projectId;
-      isSupplierQA = false;
-      res.locals.agreement_header = req.session.agreement_header;
-    }
-
-    //const baseURL = `/tenders/supplier/projects/${projectIds}/events/${eventIds}/q-and-a`;
-    //const fetchData = await TenderApi.InstanceSupplierQA().get(baseURL);
-    const data = inboxData;
-
-    const eventTypeURL = `/tenders/projects/${projectIds}`;
-
-    let getOppertunitiesData = await TenderApi.InstanceSupplierQA().get(eventTypeURL);
-    let getOppertunities = getOppertunitiesData?.data;
-
-    appendData = {
-      tender: getOppertunities.records[0].compiledRelease.tender,
-      lot: getOppertunities.records[0].compiledRelease.tender.lots[0].id,
-      data,
-      projectId: projectId,
-      //QAs: fetchData.data.QandA.length > 0 ? fetchData.data.QandA : [],
-      QAs:
-        getOppertunities.records[0].compiledRelease.tender.enquiries.length > 0
-          ? getOppertunities.records[0].compiledRelease.tender.enquiries
-          : [],
-
-      eventId: eventIds,
-      eventType: req.session.eventManagement_eventType,
-      //  eventName: projectName,
-      isSupplierQA,
-    };
-
-    res.render('viewQAList', appendData);
-  } catch (error) {
-    if (error.response.status === 401) {
-      res.redirect('/401');
-    } else if (error.response.status === 404) {
-      res.redirect('/401');
-    } else {
-      console.log('error ***************');
-      console.log(error);
-      res.redirect('/401');
-    }
   }
 };
