+++ src/main/features/digital-outcomes/path.ts
@@ -17,4 +17,5 @@ export default function (app: Application): void {
   );
   app.get(DIGITALOUTCOMES_PATHS.GET_OPPORTUNITIES_QA, DIGITALOUTCOMES_CONTROLLER.GET_OPPORTUNITIES_QA);
   app.get(DIGITALOUTCOMES_PATHS.OPPORTUNITIES_DOWNLOAD, DIGITALOUTCOMES_CONTROLLER.OPPORTUNITIES_DOWNLOAD);
+  app.get(DIGITALOUTCOMES_PATHS.OPPORTUNITY_DETAILS_DOWNLOAD, DIGITALOUTCOMES_CONTROLLER.OPPORTUNITY_DETAILS_DOWNLOAD);
 }
