+++ src/main/features/digital-outcomes/views/opportunitiesDetails.njk
@@ -13,7 +13,7 @@
             "href": "/digital-outcomes-and-specialists/opportunities"
         },
         {
-            "text": "Dorset ICS Integration and Interoperability Review & Recommendation",
+            "text": tenderer.title,
             "href": "#"
         }
     ]
@@ -95,7 +95,7 @@
                     </p>
                         </dt>
                         <dd class="govuk-summary-list__actions">
-                            <a class="govuk-link" href="#">Download opportunity details</a>
+                            <a class="govuk-link" href="/digital-outcomes-and-specialists/downloadDetails?id={{ocid}}&prId={{projectId}}&download=1">Download opportunity details</a>
                         </dd>
                     </div>
                 </dl>
@@ -166,7 +166,7 @@
                 <h2 class="govuk-heading-m govuk-!-margin-bottom-3">3. Award summary</h2>
             </div>
             <div class="govuk-grid-column-full">
-                <p class="govuk-body govuk-!-font-weight-bold govuk-!-font-size-24">{{awards.suppliers[0].name}}</p>
+                <p class="govuk-body govuk-!-font-weight-bold govuk-!-font-size-24">{{parties.name}}</p>
                 {# <p class="govuk-body govuk-body-color-warning text-strikethrough">US-DUNS-500191747</p> #}
             </div>
